all: create testutil.CleanBucket
We've copied this code around to several locations -- it's time to put
it in a shared location.

I chose testutil rather than a new package because it's only used for
tests.
