+++ healthcare/fhir_test.go
@@ -24,12 +24,10 @@ import (
 	"testing"
 	"time"
 
-	"cloud.google.com/go/storage"
 	conditionaldelete "github.com/GoogleCloudPlatform/golang-samples/healthcare/internal/fhir-resource-conditional-delete"
 	conditionalpatch "github.com/GoogleCloudPlatform/golang-samples/healthcare/internal/fhir-resource-conditional-patch"
 	conditionalupdate "github.com/GoogleCloudPlatform/golang-samples/healthcare/internal/fhir-resource-conditional-update"
 	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
-	"google.golang.org/api/iterator"
 )
 
 // TestFHIRStore runs all FHIR store tests to avoid having to
@@ -248,7 +246,8 @@ func TestFHIRStore(t *testing.T) {
 		}
 	})
 
-	testutil.Retry(t, 10, 2*time.Second, func(r *testutil.R) {
+	// Longer retry time to avoid bucket create/delete API quota issues.
+	testutil.Retry(t, 10, 10*time.Second, func(r *testutil.R) {
 		buf.Reset()
 		// Delete the bucket (if it exists) then recreate it, optimistically
 		// ignoring errors.
@@ -257,7 +256,7 @@ func TestFHIRStore(t *testing.T) {
 		// fail until they give the agent access to the bucket.
 		bucketName := tc.ProjectID + "-healthcare-test"
 		gsURIPrefix := "gs://" + bucketName + "/fhir-export/"
-		cleanBucket(tc.ProjectID, bucketName)
+		testutil.CleanBucket(context.Background(), t, tc.ProjectID, bucketName)
 
 		if err := exportFHIRResource(buf, tc.ProjectID, location, datasetID, fhirStoreID, gsURIPrefix); err != nil {
 			r.Errorf("exportFHIRResource got err: %v", err)
@@ -312,47 +311,4 @@ func TestFHIRStore(t *testing.T) {
 			r.Errorf("deleteDataset got err: %v", err)
 		}
 	})
-}
-
-func cleanBucket(projectID, bucket string) error {
-	ctx := context.Background()
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		return fmt.Errorf("storage.NewClient: %v", err)
-	}
-	deleteBucketIfExists(ctx, client, bucket)
-
-	b := client.Bucket(bucket)
-	// Now create it
-	if err := b.Create(ctx, projectID, nil); err != nil {
-		return fmt.Errorf("Bucket.Create(%q): %v", bucket, err)
-	}
-	return nil
-}
-
-func deleteBucketIfExists(ctx context.Context, client *storage.Client, bucket string) error {
-	b := client.Bucket(bucket)
-	if _, err := b.Attrs(ctx); err != nil {
-		return nil
-	}
-
-	// Delete all the elements in the already existent bucket
-	it := b.Objects(ctx, nil)
-	for {
-		attrs, err := it.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return fmt.Errorf("Bucket.Objects(%q): %v", bucket, err)
-		}
-		if err := b.Object(attrs.Name).Delete(ctx); err != nil {
-			return fmt.Errorf("Bucket(%q).Object(%q).Delete: %v", bucket, attrs.Name, err)
-		}
-	}
-	// Then delete the bucket itself
-	if err := b.Delete(ctx); err != nil {
-		return fmt.Errorf("Bucket.Delete(%q): %v", bucket, err)
-	}
-	return nil
 }
